Keep fixed-effects column in A and R
To be able to calculate standard errors for the fixed-effects more easily,
retain the fixed-effects columns in `trms`, `A`, and `R` instead of making them of width
zero.  The only modification for the PIRLS algorithm is an extra argument, `usefe` for `ranef!`.

Several methods are generalized for `MixedModel` instead of just `LinearMixedModel`
and moved from `src/pls.jl` to `src/mixedmodels.jl`