Support for bootstrapping models with a mixture of correlation structures (#577)
* Loosen type constraint on MixedModelsBootstrap.λ

Changing the type constraint on λ fixes problems when bootstrapping models
with a mixture of zerocorr (Diagonal) and non zerocorr (LowerTriangular)
RE representations. We also remove the type constraint on the parent, i.e.
 internal  storage within LowerTriangular|Diagonal so that we can better
support more types in the future. As a pleasant side effect, this makes
the type parameterization easier.

* Patch bump + NEWS